part of date text is cropped in Photocard.--> container height is increased from 18 to 20
open the favorite photo when photo card is open --> if indexOfFavoritePhoto is not null, initial item is set in scroll controller.

location of photo is not precise because size of container is overwritten --> make threshold size 20 from 30 to make it less biased.

removed year text when one of chart is not expanded as it doesn't look good